Refactored functions into a backend file.
